requirements.txt
# Zoom Attention Monitoring System - Dependencies

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Computer Vision & AI
opencv-python>=4.8.0
mediapipe>=0.10.0,<0.10.15  # Keep <0.10.15 for mp.solutions API
numpy>=1.24.0

# HTTP Requests (for Zoom OAuth)
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# WSGI Server (production)
gunicorn>=21.0.0
